updated location Int64 compatibility -> migration and code

+++ AutoTagger.Contract/Models/ILocation.cs
@@ -6,7 +6,7 @@ namespace AutoTagger.Contract.Models
 {
     public interface ILocation
     {
-        int Id { get; set; }
+        Int64 Id { get; set; }
 
         int InstaId { get; set; }
